feat: Add API and entity types for command interpretation and job applications
